+++ app/__init__.py
@@ -6,38 +6,7 @@ from flask import Flask, redirect, request, flash, url_for
 from flask_sqlalchemy import SQLAlchemy
 from flask_bcrypt import Bcrypt
 from flask_login import LoginManager
-from dataclasses import dataclass
 from flask_migrate import Migrate
-
-
-
-@dataclass
-class ItemPrice:
-    coin: int
-    metal: int
-
-
-class ShopItem:
-    def __init__(self, coin: int, metal: int, power: int, picture_name: str):
-        self.coin = coin
-        self.metal = metal
-        self.power = power
-        self.picture_name = picture_name
-
-    def price(self, amount):
-        return ItemPrice(self.coin * amount, self.metal * amount)
-
-
-SHOP_ITEMS = {
-    'pistol': ShopItem(200, 50, 1_000, 'pistol'),
-    'rifle': ShopItem(400, 150, 2_500, 'rifle'),
-    'tank': ShopItem(800, 330, 6_000, 'tank'),
-    'missile_1': ShopItem(1_600, 700, 12_000, 'missile_1'),
-    'missile_2': ShopItem(4_000, 1_800, 25_000, 'missile_2'),
-    'missile_3': ShopItem(10_600, 3_700, 50_000, 'missile_3'),
-    'jet': ShopItem(35_000, 12_000, 150_000, 'jet'),
-}
-
 
 
 basedir = os.path.abspath(os.path.dirname(__file__))
